requirements.txt
pypdf
google.generativeai
chromadb
flask
flask_cors
gunicorn